Another fsync optimization in the negotiator. #2342
This makes AddMatch() and RemoveMatch() nondurable in order to solve
severe performance problems at large scale in Accountant::CheckMatches().
This change does not alter the consistency model of match records
in the accounting database: the contents of the database are not and
never were authoritative; the match records are updated in each negotiation
cycle to conform to the information published by the startds.  Therefore,
there is no need for these records to be synced to disk.