Fix scheduler universe jobs with spooled files. #2396
If a job has its files spooled, the Cmd attribute will be the basename
of the executable. Scheduler::start_sched_universe_job() expected Cmd
to contain the full path, so starting the job would fail.

Now, if Cmd isn't an absolute path, the schedd will prepend the Iwd
value to create one.