Disallow hold=true when spooling job files. #2398
hold=true conflicts with job-file spooling. Both want to set a
HoldReason. If the wrong HoldReason is set, the schedd doesn't modify
the paths in the job ad to refer to the spool area. Also, the job ends
up in the idle state once spooling completes.
Now, condor_submit fails with an error if the user tries to use these
together.