Patch debian package script to deal with externals libraries. #2389
When calculating shared library dependencies, dpkg-shlibdeps doesn't
account for libraries not being in their final installation location and
thus not being found via rpath. So we set LD_LIBRARY_PATH to
<package dir>/usr/lib/condor when calling dpkg-shlibdeps. We only do
this for non-proper builds.
We also re-add the --ignore-missing-info option, since the libraries
from our externals don't have explicit dependency info and we no longer
build on debian 4.0 (the reason we removed this option).