Changes from code review of gittrac #2057, part 3
Nathan,

* It looks like if the PRE script fails but the POST script is run,
the POST script will get 0 for the $RETURN value.  That doesn't seem
like a good idea to me.  I'm thinking that maybe we should set it to
something like
const int Dag::DAG_ERROR_JOB_SKIPPED = -1004

* In job.h, GetPreStatus() *really* needs to be documented, because it's
not just returning the status the PRE script returned, but messing with
it if there's a PRE_SKIP value.

Kent