===GT=== #2337 ===GT:Fixed=== #2337 The purpose of this commit is to allow standard universe's file I/O, both remote and local, to read and write files larger than 2GB.
This is the exact problem:

In many places in the stduniv code, the sizes of files or the integer
positions of files (used by lseek, etc) were recorded as ints. This
is often a narrower type than off_t, which is what *should* be used to
record that information. The type narrowing limited the reading/writing
of files a condor_compiled application could perform to 2GB.

This is how it was fixed:

The stub generator was primarily responsible for the severe misuse of
the return type of all system calls. The stub generator, before this fix,
hard coded the return type of every single system call to be an integer,
regardless if it was a pointer or another differently sized entity.
I fixed the stub generator to emit a senders/receivers/switches file that
is fully type correct. Now, each system call's return value is what it
is supposed to be and cedar (which already had this functionality) knows
what to code/decode on the wire. Some minor features had to be added
to the stub gen language to handle sorted edge cases of type handling
(like handling the void return type of some system calls, etc).

The other major culprit in this problem was that the CondorFileTable
object heirarchy also used the int type in calculations and storage of
sizes/offsets into files. These uses were subtle and hard to find. I
hope I got them all.

I wrote a test, called the job_ckpt_lfs_std test. It is used to create a
5GB file, populate it with a randomly generated, but repeatable, set of
unsigned chars. The file is written, then read, and the numbers better
match in the order that it expects.

I also juiced up some of the dependency links in the Cmake system to be
better about stuff in the stduniv. There is still a lot of work to be
done in it to make it usable when doing actual stduniv development. :(

Caveats:

It turns out that the size of off_t is a funny thing. On 32-bit machines
off_t is normally a 4 byte quantity, but the preprocessor defines
_LARGEFILE_SOURCE and _FILE_OFFSET_BITS=64 will change it to an 8 byte
quantity. This is very problematic for us. Condor's stduniv library
is compiled WITHOUT those two preprocessor defines, so Condor's code
assumes a 4 byte off_t. If the application object files were compiled with
those two defines and then condor_compiled with the stduniv libraries,
there will be a size mismatch between the off_t that the application
has (8 bytes) and the size Condor uses (4 bytes).

It seems on 64-bit machines, those two flags don't affect the off_t size.

Given this problem, here is a chart of what works:

----
I will call the compilations of Condor that do not use the two #defines
as 32CN 64CN. This is the only way Condor is compiled today.

Now, an application can be compiled 32-bit or 64 bit, and either have
the above constants defined for it or not. I will call these 32AN 32AY
64AN 64AY.

Then here is a chart which describes if an application doing 64-bit file
offsets with remote i/o will work with a specific type of Condor:

Application Type     Condor Shadow     64-bit File Offsets Supported
32AN                 32CN              NO
32AN                 64CN              NO
32AY                 32CN              NO
32AY                 64CN              MAYBE
64AN                 32CN              NO
64AN                 64CN              YES
64AY                 32CN              NO
64AY                 64CN              YES
----

Recommendations:

The upshot is that 64-bit shadows and 64-bit applications will be able
to read/write files larger than 2GB. Anything else will not work or is
currently not supported.

Oddities:

It seems that the getdirentries/getdents system call was very fragile in its
implementation. Before this commit, they worked in remote i/o mode, but
failed in local mode. After this commit, they will fail in both modes.
I need to debug this and it is a new ticket.