Insert PRE_SCRIPT_RETURN as a SCRIPT variable
===GT=== #2057
The _retValScript was already a member of the Script class, but
was not being used.  So I appropriated it to use as the return
value from the PRE script.  This can be passed to the POST
script by making reference to PRE_SCRIPT_RETURN in the DAGman
file.

Document the use of the PRE_SCRIPT_RETURN variable in the manual.