Fix handling of stdout/err containing paths when not using file transfer. #1811
When should_transfer_files is set to YES or IF_NEEDED and stdout/err
contain any path information, condor_submit renames them to
_condor_stdout/err and uses transfer_output_remaps to specify the actual
names. However, if no file transfer is used, the remaps aren't applied
and the output gets written to _condor_stdout/err instead of the
locations specified by the uesr.

This hits vanilla universe when IF_NEEDED is used and the starter
decides it doesn't need file transfer. It also hits scheduler and local
universe.

Now, condor_submit only does this renaming when spooling files to the
schedd. When not spooling, the shadow and starter do the renaming, but
only when actually using file transfer. When using all new daemons, the
files always end up where the user expects. When mixing new and old
daemons, the behavior will be no worse than before.

===VersionHistory:Complete===