===GT=== #2467 This is the first part of the fix for the broken getdirentries() problem. Also, it turns out that this problem affected other system calls and Condor 7.7.1 is a little messed up over it.
The problem was that when I refactored the stub generator to be type
correct for return values of system calls, I missed the handling of the
'code_bytes_bool' expression in the syscall.tmpl file. That expression
names the variable symbol in the generated source which holds a length
quantity for the coding of a binary array. I had changed some aspects of
the stub generator to hold the returned value in a symbol called "ret_val".
However, "rval" was a symbol that was still valid in the scope of all
of the calls in receivers.cpp. So I ended up choosing the wrong symbol and
only wrote a tiny fraction of the bytes in an array back to the job. This
fix makes ALL handling of the return valus of system calls use the "ret_val"
symbol so it shouldn't happen anymore. I never saw the error from the compiler
because it turned out, unluckily, that "rval" and "ret_val" were of the
same type in the calls that were failing.