Fix version check in shadow for transfer IF_NEEDED bugfix. #1811
If the shadow is talking to an old starter and file transfer may be
used, it needs to rewrite the job ad it sends to the starter so that
stdout/err get written to the correct location.

I tried to get the starter version from the FileTransfer object, but it
doesn't learn the starter's version until later. Instead, we can use the
peer version from the ReliSock of the connection to the starter.