Remove compatibility code for Submitter/Scheduler ads in the collector. #1584
Finish transition to always label submitter ads as "Submitter" on the
wire instead of as "Scheduler".
We remove code in the collector that forces the ad type to "Submitter"
for submitter ads and queries for them. We assume all clients use the
correct name now.
We also remove code in the collector that always includes the NumUser
attribute in scheduler ads sent in response to a query.
===VersionHistory:None===