Turn libcondor_utils into a dynamic library on mac. #2132
The environ variable is not directly accessible by dynamic libraries on
the mac, but there is a function available to get its location. As a
result, Condor now has the function GetEnviron() for utility code that
needs to directly access environ.  On the mac, it calls the system
accessor function. Everywhere else, it returns environ.