Remove make test from openssl build
Theres a bug in the tests where it occasionally makes a random
seed that it can't parse.  It fails roughtly one build in 20
but usually right before an important release.  We're not
going to fix the test, and the openssl people aren't going
to fix the test, so let's live dangerously and not run the
test suite on every build.