Submit event not written for spooled jobs when running as root #2569
When a job's files are first spooled to the schedd, they are owned by
the condor user. Since the user log is usually written to as the user,
logging of the submit event fails when files are spooled.

To fix this, we change Scheduler::InitializeUserLog() to try opening the
user log as condor if opening it as the user fails and the file is in
the spool directory.
===VersionHistory:Complete===