Fix race condition when putting running jobs on hold. #2577
After running condor_hold, a job update received by the
shadow after requesting a deactivation but before
the job was evicted would cause the job to change from
the held state to the running state.  Then when the job
was finally evicted, the job would leave the queue rather
than being put on hold.  A spurious unsuspend event
was also logged in the userlog.