Make procapi on macos properly detect non-existent pids. #2594
If ProcAPI::getProcInfo() is called for a non-existent pid on macos, it
uses random data from the heap as the pid's stats. This is because when
sysctl() is called on an invalid pid, it returns success and indicates
that no data was written in the provided struct. getProcInfoRaw() was
only checking the return value. Now, it checks for the amount of data
written as well.
===VersionHistory:Complete===