Disable classad conversion function in classad analyis code. #2603
The conversion function toNewClassAd() trips up on string attributes
that end in a backslash. Thus, condor_q -analyze fails when a job or
machine ad contains such an attribute. Since we now use new classads
everywhere, this conversion function is unnecessary. Removing calls to
it fixes this bug and improves performance.
===VersionHistory:Complete===