Initial support for absent ads in the collector. #2608
Added knob ABSENT_REQUIREMENTS which, if defined, is evaluated
whenever a startd ad in the collector would expire. If it evals
to True, the ad is stored in the offline log just like offline ads.
Queries sent to the collector are modified to not include
ads where ABSENT=True unless the query explitly references the
ABSENT attribute.