Cleaned up semantics of Env object.  Fixed failing unit test. #2614
Now using Import() in place of MergeFrom() in several places,
because the former is intended for the case where one wishes
to silently skip over indigestible entries in the environment.
However, the semantics of MergeFrom() are now at least less
likely to cause subtle bugs when it encounters such cases.

Running out of memory when inserting into the environment now
blows up in an assert rather than returning an error code
(which is often ignored).