Schedd's job evaluation timer can become unregistered. #2647
A reconfig delivered at the wrong moment can cause the schedd's
Scheduler::timeout() timer not be re-registered. This leads to the
schedd never starting idle jobs or sending ads to the collector. Only a
reconfig or restart can restore normal operation.

To fix this, we now never cancel the timer and always register it as a
periodic timer (instead of as a one-shot timer).
===VersionHistory:Complete===