Always build collector offline plugin regardless of HAVE_HIBERNATION. #2608
Previously the collector invoked offline plugin methods conditional
on the HAVE_HIBERNATION compilation option. But now that this
plugin is also used for ABSENT_REQUIREMENTS, invoke these
methods unconditionally. Note that all these methods result in
a no-op in the collector at runtime if the OFFLINE_LOG is
not defined (which is the default).