Add remote_gahp script for running the blahp remotely. #2421
remote_gahp is a small script that allows the gridmanager to run the
blahp on a remote system via ssh. This allows submission to a
pbs/lsf/sge cluster on a remote machine without using Condor-C.

The blahp must be pre-installed on the remote machine and major
features, like file transfer, are not supported as present.