Disable use of LastJobLeaseRenewalFailed in CalculateJobLease(). #2455
When a job lease update fails, this attribute causes subsequent update
attempts to use the same expiration time, even if that time has since
elapsed. For cream grid jobs, this causes all jobs going to a given
resource to fail, as the job lease is shared.

Also add a sanity check for the lease duration value.