Fix logging for c-gahp. #2132
When introducing the new set_mySubSystem() call, I put it in the wrong
location in the c-gahp. Consequently, the c-gahp thought its subsystem
was TOOL. When Condor was started as root, the c-gahp would fail
immediately because it couldn't write to $(LOG)/TOOLLog.

No version history for this fix, as the bug didn't make it into a release.