Fix submission of grid jobs via soap. #2568
When a grid universe job is submitted via soap, it's spool directory
isn't chowned to the job owner. The gridmanager and its children access
job files as the job owner, so any attempts to write in the spool
directory fail.

To fix this, we add an explicit call to aboutToSpawnJobHandler() in the
soap submission code path for grid universe jobs.
===VersionHistory:Complete===