Finish hold-claim for DAGman
===GT=== #2673
===VersionHistory===

Resolves the following comments from Kent:

> Condor DAGman will hold onto a claim for a short while before releasing it.
> The default hold time is 20 seconds, but this can be changed in the DAGman
> configuration file. The change is effected by using the
> DAGMAN_HOLD_CLAIM_TIME
> configuration variable.
> ---

Okay, here are my comments:

DAGMAN_HOLD_CLAIM_TIME is not documented in the "configuration
variable..." section of the version history.  (One other note on this -- I
thought that the items about new config variables were supposed to have a
link to the place in the manual where the config variable is actually
fully defined.  But most of them seem not to.  I don't know if that's a
problem, or if I'm just wrong in expecting that.)

In the documentation of DAGMAN_HOLD_CLAIM_TIME in the config section, it
would be great to make the reference to KeepClaimIdle be a link.
(Note: I don't see how this is done in the manual --- nwp)

I think we should also mention this in the actual DAGMan section of the
manual, maybe with some higher-level explanation of how/why you'd want to
use it.

In the test, KeepClaimIdle is undefined for node C.  Is that because it's
the last node in the DAG?  If so, it would be good to put a comment about
that in the test DAG and/or .run file, and the documentation.

I guess we don't need 'Condor::DebugLevel 10' in the test script anymore.

hold_claim in condor_submit() (and prohibitMultiJobs, for that matter) is
not documented in submit.h.

Other than that, it looks good to me!  Sorry it took so long for me to get
around to reviewing this.

Kent