Warning fixes.
Fix types of variables used for tracking input and output from write; should be ssize_t (output) and size_t (input).  Removes signed/unsigned comparison warning.

Wrapped write and read with a version that exits with an error on any errors.  previously we were ignoring errors, which would likely cause further problems.  A "proper" fix of dealing with errors is probably not worth the time for the hopefully soon to be retired checkpoint server.