Make gridmanager more tolerant of nordugrid ldap responses. #2888
When querying a nordugrid ldap server for job statuses, the gridmanager
expected all requested attributes to be present and in a certain order
in each record. Otherwise, it ASSERT()ed. We've now seen servers that
don't follow these expectations. Now, the gridmanager doesn't care about
ordering of attributes and ignores records with missing attributes.
===VersionHistory:Complete===