Fixed an infinite loop in the gridmanager for gt2 jobs. #2916
If the jobmanager is listening on a different port than what's in
GridJobId, the GlobusJob object will bounce between states GM_REGISTER
and GM_RESTART, always attempting to contact the jobmanager using the
incorrect port.
Now, when a restart fails with error OLD_JM_ALIVE, the gridmanager
updates GridJobId with the job contact string returned by the restart
command (which has the correct port).
===VersionHistory:Complete===