Re-add Condor version and platform strings to binaries #2932
The ident-style strings for CondorVersion and CondorPlatform disappeared
from the Condor binaries when we switched to dynamically linking the
Condor utility libraries. Now, they're included again.

We do this by adding condor_version.cpp to the list of source files for
any binary built using the condor_daemon() and condor_exe() cmake
macros.