Make shadow understand new UserLog format
This is compatible with the old Userlog format, unless ';' is used as a
separator.  There is probably some system that I don't know about uses
semicolons as a regular part of the filename

The idea here is that if the UserLog is specified in the JobAd as

UserLog = "bob;bill"

then the shadow will write a two copies of each event: one into the
"bob" file and one into the "bill" file.

I think if the event log is enabled, we get a copy of each event for
each userlog entry. This is a bad thing; I have not tested it yet.

===GT=== #2807