Use new macro capability in DAGman
DAGman now requests the schedd and shadow to write each event to the
default log, i.e., <DAG file>.nodes.log, along with whatever log is
specified by the user.

DAGman only looks in the default log for events.

DAGman only writes POST Script TERMINATE events to its default log, not
to the user log.

DAGman specifies that the default user log is in plain text format, not
XML format.

This allows the user to have his own templated UserLog file, as
requested in #2428.

This does not completely implement the solution specified in #2807, as
we still write to a log which is also being written to by the schedd and
shadow.  However, it does fix the fd-throttling issue in #659, as there
is only one userlog that condor_dagman watches for events.

===GT=== #2807