Made semantics of GSI_DAEMON_NAME more backward compatible. #1605
When GSI_DAEMON_NAME is defined, only certs that match it are accepted.
This changes the behavior introduced in my previous patch, which
accepted certs matching GSI_DAEMON_NAME _or_ certs passing the
hostname check.  Now the addition of hostname checks is strictly
more or equally restrictive compared to the old behavior,
given the same configuration.  I think that is a safer change.