Fix cmake compiler choices on newer macs.
By default, cmake 2.8.x will pick gcc and c++ as the C and C++ compilers
on linux and mac. With the newest mac dev tools, these point to
completely different compilers (llvm gnu and clang, respecitively).
To pick a consistent set of compilers, we'll set the environment
variables CC and CXX.