Fix ups based on feedback #2807
This implements suggestions based on design documentation review

Remove capability to have a vector of arbitrary length fed to the
schedd; that is a recipe for a DoS. Add the "dagman_log" field; old
versions of the schedd will ignore it and thus it will only affect the
dagman user. Also, UserLog will be the same as it has always been---a
filename.

Add "DAGManNodesLog" as an attribute name.

Document this in the manual.