Input files no longer transferred as output files. #2406
If a job's input files were stored in the spool directory, then the
starter would transfer all input files along with the output files after
job completion. This is because the FileTransfer object in the shadow
considered all files in the spool as intermediate files created in a
previous partial-execution of the job.
Now, when the job's input files are spooled, none of the files there are
considered intermediate files by the shadow's FileTransfer object.
It's ok to treat real intermediate files this way because files created
in previous runs are already in the job's Iwd (the spool directory) and
so don't need to be transferred, and any files modified by the new run
of the job will be transferred as output files.
===VersionHistory:Complete===