Fix unparsing of non-standard ClassAd attribute names. #3099
ClassAd attribute names can contain any character. If they contain
characters other than alphanumerics and underscore, the name must be
enclosed in single quotes. The native unparser didn't do the proper
quoting if the first character of the attribute name is something other
than an alphabet letter or underscore. That is now fixed.