Remove ClassAd functions from ClassAd interface: EvaluateAttrList and EvaluateAttrClassAd. #3082
These functions are not compatible with dynamically allocated values, because the
evaluated Value goes out of scope.  A shared_ptr could be used instead, but this has
performance implications that depend on internal implementation details of ClassAds,
so I think it is better not to offer these methods until we find a case where
they are actually needed.