Disable classad caching by default, add api to enable on request. #3127
The classad caching code makes classads much less thread-safe. Threads
using different classads now operate on shared data (the cache and the
shared ExprTrees). This is causing frequent crashes in the blahp.

The short-term solution is to disable caching by default (regardless of
the parameter to ClassAd::Insert()) and add a function to enable it
process-wide on demand.