(#3139) condor_q displays description if one is defined
When a given job has a description, it will print it instead of the command in the CMD column.

The given example shows the output (without the -wide option):

 ID      OWNER            SUBMITTED     RUN_TIME ST PRI SIZE CMD
8932.0   santos-g        7/20 16:26   0+00:00:00 I  0   97.7 python python_job.
8933.0   santos-g        7/20 16:28   0+00:00:00 I  0   97.7 (Interactive from )
8936.0   santos-g        7/20 17:29   0+00:00:00 I  0   97.7 (hey this is a rea)

Job (#8932.0) has no description.
Job (#8933.0) is interactive with no description, so a default description is given to it and shown there.
Job (#8936.0) was given a description.

And below is the output with -wide:

 ID      OWNER            SUBMITTED     RUN_TIME ST PRI SIZE CMD
8932.0   santos-g        7/20 16:26   0+00:00:00 I  0   97.7 python python_job.py 0
8933.0   santos-g        7/20 16:28   0+00:00:00 I  0   97.7 (Interactive from condor-student1.cs.wisc.edu)
8936.0   santos-g        7/20 17:29   0+00:00:00 I  0   97.7 (hey this is a really looong description!)