Maximize use of shared fs in the blahp #3195
The blahp has code to determine when it can rely on a shared filesystem
and when it has to employ explicit file transfer requests in the
underlying batch system. This is done in the function
bls_test_shared_dir() in blah_common_submit_functions.sh. This function
always says relative pathnames cannot be accessed via a shared
filesystem. As a result, if the ad given to the blahp by Condor contains
a relative path, explicit file transfers are used even when they aren't
necessary. This is a problem for SGE and some flavors of PBS, where file
transfers by the batch system aren't supported.

To fix this, we can change bls_setup_all_files() in
blah_common_submit_functions.sh to always use absolute paths when
calling bls_test_shared_dir(). Relative filenames are relative to the
job's Iwd.