condor_q misses job ads when schedd version unknown #3206
When using the old schedd query interface in
CondorQ::fetchQueueFromHostAndProcess(), the unparsed constraint string
can become corrupted if the caller-provided function pointer calls
ExprTreeToString(). As a result, we may not get all of the desired ads
from the schedd.

To fix this, we copy the constraint string returned by
ExprTreeToString() in CondorQ::fetchQueueFromHostAndProcess().