New submit attribute use_x509userproxy. #3025
Set to True to indicate that an X.509 user proxy is required for the
job. This attribute allows the use of a proxy in one of the standard
locations without having to provide the path to the file using the
x509userproxy attribute.

There is no new corresponding job ad attribute. This submit attribute
merely ensures that a valid proxy is available and that X509UserProxy
and friends are set appropriately in the job ad.