BOSCO: Better replicate remote environment in remote_gahp.
Use -l instead of -i when invoking bash remotely. The set of
initilization scripts is different for a login shell vs an interactive
shell. The login setup will usually be closer to what a user sees at
their command prompt.

We've seen one machine where the batch system's bin directory was added
to the PATH in .bash_profile. Thus, a login shell has it in the PATH but
a non-login interactive shell does not.