Fix syntax error in Cream job ad sent to gahp.
Due to our switch to the real new ClassAds parser/unparser, the job ad
the gridmanager sends to the cream_gahp was missing a semicolon. The
Cream server was thus unable to parse the ad. This is now fixed.

Note: This was not a bug in the ClassAd unparsing code, but an
assumption in the gridmanager about trailing semicolons after the last
attribute in an unparsed ad.