Make grid-type batch work on systems with glite installed. #3269
The gridmanager sets GLITE_LOCATION in the environment so that the blahp
we distribute can find its files. If a system has glite installed, the
login scripts will set GLITE_LOCATION to that location. The blahp's
submit scripts run bash with -l, causing the login scripts to be run,
wiping out our value of GLITE_LOCATION.

So we add a patch to remove the -l in the submit scripts. Nothing else
in the login scripts should affect job submission.