Scrub ft-gahp sandbox ids of problem characters. #3294
We use COLLECTOR_HOST as part of a sandbox id for the ft-gahp. This
sandbox id becomes the name of a directory, which is given to the blahp
and the batch system behind it. Certain characters that can appear in
COLLECTOR_HOST will cause problems when part of a file path, so we need
to scrub them from the sandbox id.