Don't leave jobs in hold when condor_submit dies #3200
When doing a condor_submit with the -spool option, condor_submit first
submits all the jobs on hold in one transaction, and then it starts a
second transaction to transmit the input files.  If there is an error
sending the input files, condor_submit will exit with a nonzero exit
status (this is good), but a pile of jobs will be left behind in the
queue still sitting on hold (this is bad).  If condor_submit returns
failure, the job queue should be left in a state as if nothing happened
- aka the entire submission should be rolled back.

If the staging of the input files failed, schedd in method
Scheduler::spoolJobFilesReaper() should remove all the jobs involved
(this same method is already releasing all the involved jobs from hold
if the transfer succeeds).

In addition, in the rare event that condor_submit dies right after
submitting the jobs but before it starts transferring the input files,
the schedd should periodically remove (garbage collect) jobs that are in
hold state due to file transfer for a long period of time, or for a
short period of time if there are no file worker threads active (i.e.
spoolJobFileWorkers hash table is empty).

To keep Condor-C happy, in Scheduler::spoolJobFiles, alongside the code
where we look at ATTR_STAGE_IN_FINISH to be certain we do not allow
stagein to happen more than once, we should also confirm that the job is
in held state for stagein reasons - this will keep Condor-C from
attempting to restage files to a job in X state

The above text is in the ticket, by Todd T.