Blahp shouldn't alter LD_LIBRARY_PATH. #3317
The blahp adds several directories to LD_LIBRARY_PATH. The way we build
it, this is unnecessary. But we've seen some instances where it screws
up some of the command line tools of the local batch system when called
by the blahp. So we're disabling the modifications to LD_LIBRARY_PATH.