Always set GLOBUS_LOCATION for the blahp. #3015
If GLOBUS_LOCATION isn't set in the blahp's environment, then it tries
to invoke proxy-handling tools under /opt/globus, which is almost never
going to be the correct location anymore. /usr is a better default.

Now, GLOBUS_LOCATAION for the blahp can be set via the config file
(using the same name). If it's not set in the config file or in the
environment of the gridmanager, then it's set to '/usr' in the blahp's
environment.