Fix frequent redelegation of Condor-C job proxy. #3395
If HTCondor is configured not to limit the lifetime of a delegated proxy
(param DELEGATE_JOB_GSI_CREDENTIALS_LIFETIME or job attribute
DelegateJobGSICredentialsLifetime set to 0), then the gridmanager will
frequently redelegate the proxy of Condor-C jobs to the remote schedd.
The gridmanager's constant redelegation is only limited by
GRIDMANAGER_PROXY_REFRESH_INTERVAL, which defaults to 10 minutes.

Add a check in the gridmanager for the don't-limit-delegation-lifetime
case.