Recover from PRE_SKIP #2966
This is for recovery in DAGman.  If we need to enter recovery mode, then
we need to have the events in the DAG so that recovery can proceed.

Correctly read the PreSkipEvent from the UserLog.

Simplify the code by just saying that PRESKIP trigger sends a job
straight to STATUS_DONE

Do not increment the submit counter if we are a PRESKIP job.

Write a test

Terminate a job, and include a Node with only a Pre Script. This touches
a different code path than nodes with a post script.

Include a version history comment

The test script delete the lock file
(job_dagman_recovery_preskip.dag.lock) if it exists, just in case we
have one hanging around for some reason.