condor_submit should allow schedd to set Owner for remote submits. #3370
When spooling files with condor_submit, set Owner=Undefined in the job
ad. This will tell the schedd to set Owner based on the user name that
came from the authentication step.
When spooling files, there's no need or expectation for the job owner in
the schedd to match the user running condor_submit.