Implement a volatile update mechanism for chirp. #3353
This is designed to be a scalable mechanism to send monitoring
updates from the jobs.

We add a new command to chirp, set_job_attr_volatile.  When used,
the starter will send the attribute update via the resource update
mechanism (same message that does CPU time, memory usage, etc).

This variant of set_job_attr uses far less resources on the schedd
as the updates ride along with the existing mechanism and may not
necessarily be a durable update.  The downside is the update may
not get to the schedd and be lost.

This does not force an update immediately - these are sent along
with the next update from the starter.  If you call this multiple
times for the same attribute before the update is sent, the last
value is sent.