Add support for RequestMemory to the blahp. #3398
Allow users of the blahp to specify the amount of memory required by the
job, using the same attribute as for HTCondor job ads. This number (in
megabytes) is passed to the submit scripts via the new -m parameter. The
PBS and Condor scripts now use it in the job submission. The other
submit scripts don't use it yet.