Try to fix Windows7 in Batlab.  #3584
Boost chokes if you explicitly ask it to enable python in the
bootstrap but Python.h is not present.  This patch enables
--with-python in the bootstrap only if both the interpretter and
libraries are found.  On Windows7, the headers are missing.

Trying to not have to conditionally disable python on Win7 just
in case someone else out there wants to get it going.