Fix authenticated name reporting in CEDAR for GSI. #3612
ReliSock::getAuthenticatedName() was returning an empty string when the
client authenticated with a GSI credential with no VOMS attributes. This
was due to Condor_Auth_X509::getFQAN() returning an empty string in this
case, whereas its caller was expecting to to return NULL.
Changed Condor_Auth_X509::getFQAN() to behave as the caller expects.