Indicate an error when we cannot submit a nested DAG #3623
Hide the CondorID. When looking for bugs in #3623, I tried to find all
the places we changed the ID, but it was difficult. It turned out to be
a false path as far as the bug goes, but seems like a good thing to do
anyway.

Remove the swap_priorities hack.  Use RAII instead to let the compiler
fix up priorities.

Add a test. Add to CMake test list.

Add version history.