Don't fork-bomb the system when restarting the procd. #3618
When attempting to restart the procd, cap the number of attempts.
Otherwise, we can end up creating an unlimited number of procds until we
reach a system limit on the number of processes.
This is a very simple fix where we suicide on failure, which is what
would happen at the end of the fork-bomb. More graceful error-handling
may be called for in a future version.