Clean out gsoap symbols from global table.  #3625.
HTCondor daemons previously exported various statically-compiled
and gsoap-generated symbols globally to the dynamic linker.

If a plugin is loaded which internally uses gsoap (such as happens
with Globus GSI auth plugins that eventually callout to GUMS), then
the dynamic loader will link against some symbols from HTCondor and
some symbols from the plugin.  Typically, this results in a segfault
of the daemon.

This commit changes a few things:
 - Put all generated soap functions into a C++ namespace that we can.
 - Any function that is required to be global by gsoap itself is
   now hidden by the linker.
 - Instruct the Linux linker to not re-export symbols from the
   static libgsoapssl++.

With this commit, I have verified that GSI callouts no longer blow
up for the HTCondor-CE on the OSG.