Retain a few global symbols for SOAP-enabled daemons.  #3625
When we removed all global symbols from the SOAP-enabled daemons,
this broke SOAP integration.  When the dynamic linker loads
libcondor_utils, it will see weak symbols from there (which are
no-ops) and prefer them over the functioning ones in the daemon.

So, we export the strong symbols necessary for gsoap and only
those symbols from the daemon.